backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
sqlalchemy[asyncio]>=2.0.0
aiosqlite>=0.20.0
python-multipart>=0.0.12
pydantic-settings>=2.6.0
openai>=1.50.0
pymupdf>=1.25.0
python-docx>=1.1.0
jinja2>=3.1.0
chardet>=5.0.0
openpyxl>=3.1.0
snowballstemmer>=2.2.0
pytest>=8.0.0
pytest-asyncio>=0.24.0
httpx>=0.27.0
